7.Lecture-Regular_Expressions/problem_set/working.py
# File Name: working.py

# Description: This script defines a 'convert' function that converts time ranges from a specific format to a standardized format.
# The script also includes a 'main' function to interactively input time ranges and display the converted format.

# Import necessary modules: re for regular expressions.

import re

# Define the main function.
def main():
    # Get the user's input for the time range and display the converted format using the 'convert' function.
    print(convert(input("Hours: ")))

# Define the 'convert' function to convert time ranges to a standardized format.
def convert(s):
    # Use regular expressions to validate and parse the input time range.
    if matches := re.search(
        r"^(\d{1,2})(:[0-59]{2})? (AM|PM) to (\d{1,2})(:[0-59]{2})? (AM|PM)$", s
    ):
        # Extract individual components (hours, minutes, and noon) from the matched groups.
        hr1, min1, noon1, hr2, min2, noon2 = matches.groups()

        # Check if the hours are greater than 12, which is not a valid format.
        if int(hr1) > 12 or int(hr2) > 12:
            raise ValueError

        # Format and convert the times to a standardized format.
        time1 = time_format(hr1, min1, noon1)
        time2 = time_format(hr2, min2, noon2)

        # Return the converted time range in the standardized format.
        return time1 + ' to ' + time2
    else:
        # If the input does not match the expected format, raise a ValueError.
        raise ValueError

# Define the 'time_format' function to format hours and minutes.
def time_format(hrs, mint, noon):
    # Convert hours to an integer.
    hrs = int(hrs)

    # Adjust hours based on AM/PM and handle noon/midnight cases.
    if noon == "PM" and hrs != 12:
        hrs += 12
    elif noon == "AM" and hrs == 12:
        hrs = hrs - 12

    # Format the time string based on whether minutes are provided or not.
    if mint is None:
        time = f"{hrs:02}:00"
    else:
        time = f"{hrs:02}{mint}"

    return time

# Check if the script is run as the main program.
if __name__ == "__main__":
    # If it is, call the 'main' function to interactively input time ranges and display the converted format.
    main()

# Explanation:

# Description: This script defines a 'convert' function that is used to convert time ranges from a specific format to a standardized format.
# The script also includes a 'main' function for interactive use, allowing the user to input time ranges and see the converted format.

# We start by importing the 're' module for regular expressions.

# The script defines three functions:

# - 'main': This function is the entry point of the script. It gets the user's input for the time range and displays the converted format
#   using the 'convert' function.

# - 'convert': This function is responsible for validating and converting time ranges. It uses regular expressions to validate the input
#   against a specific pattern. If the input matches the pattern, it extracts the individual components (hours, minutes, and noon)
#   and checks if the hours are greater than 12 (not a valid format). If everything is valid, it calls the 'time_format' function to
#   format and convert the times to a standardized format and returns the converted time range. If the input does not match the expected
#   format, it raises a 'ValueError.'

# - 'time_format': This function formats hours and minutes based on whether the input includes minutes or not. It adjusts hours for AM/PM
#   and handles cases involving noon/midnight.

# The script also checks whether it is executed as the main program. If it is, it initiates the interactive process by calling the 'main'
# function, allowing the user to input time ranges and see the converted format.

# These functions ensure that the 'convert' function accurately converts time ranges from the specified format to a standardized format,
# handling validation and formatting appropriately.
